+++ funcx_endpoint/funcx_endpoint/engines/__init__.py
@@ -1,0 +1,4 @@
+from funcx_endpoint.engines.globus_compute import GlobusComputeEngine
+from funcx_endpoint.engines.process_pool import ProcessPoolEngine
+
+__all__ = ["ProcessPoolEngine", "GlobusComputeEngine"]
