+++ funcx_endpoint/funcx_endpoint/executors/__init__.py
@@ -1,5 +1,3 @@
-from funcx_endpoint.executors.gc_executor.gc_executor import GCExecutor
 from funcx_endpoint.executors.high_throughput.executor import HighThroughputExecutor
-from funcx_endpoint.executors.process_pool.process_pool import ProcessPoolExecutor
 
-__all__ = ["HighThroughputExecutor", "ProcessPoolExecutor", "GCExecutor"]
+__all__ = ["HighThroughputExecutor"]
